requirements.txt
flask
line-bot-sdk
gunicorn
pytrends
feedparser
beautifulsoup4
requests
